change reset-password structure
