+++ src/App.tsx
@@ -11,8 +11,8 @@ import useInitializeApp from './hook/custom/useInitializeApp';
 const MapLayout = lazy(() => import('./Components/Map/MapLayout'));
 const LoginRoute = lazy(() => import('./routes/login'));
 const SignUpRoute = lazy(() => import('./routes/sign-up'));
+const ResetPasswordRoute = lazy(() => import('./routes/reset-password'));
 const Unauthorized = lazy(() => import('./Components/Unauthorized/Unauthorized'));
-const ResetPasswordView = lazy(() => import('./Components/ResetPassword/ResetPasswordView'));
 const ConfirmPasswordView = lazy(() => import('./Components/ConfirmPassword/ConfirmPasswordView'));
 const DetailedPageView = lazy(() => import('./Components/DetailedPage/DetailedPageView'));
 const ProfileView = lazy(() => import('./Components/Profile/ProfileView'));
@@ -33,7 +33,7 @@ const App = () => {
         <Route path={`/login`} component={LoginRoute} />
         <Route path={`/sign-up`} component={SignUpRoute} />
         <Route path={'/404'} component={Unauthorized} />
-        <Route path={`/reset-password`} component={ResetPasswordView} />
+        <Route path={`/reset-password`} component={ResetPasswordRoute} />
         <Route path={`/confirm-password`} component={ConfirmPasswordView} />
         <Route path={`/detailed-page`} component={DetailedPageView} />
         <Route exact path="/" render={() => (
