+++ src/routes/reset-password/components/ResetPasswordProvider.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { GoogleReCaptchaProvider } from 'react-google-recaptcha-v3';
+import ResetPasswordLayout from './ResetPasswordLayout';
+
+const ResetPasswordProvider = () => {
+  return (
+    <GoogleReCaptchaProvider reCaptchaKey={process.env.REACT_APP_CAPTCHA}>
+      <ResetPasswordLayout />
+    </GoogleReCaptchaProvider>
+  )
+};
+
+export default ResetPasswordProvider;
