+++ src/routes/reset-password/index.ts
@@ -1,0 +1,3 @@
+import ResetPasswordProvider from './components/ResetPasswordProvider';
+
+export default ResetPasswordProvider;
